backend/requirements.txt
Django>=5,<6
djangorestframework>=3.15,<4
django-filter>=24,<25
djangorestframework-simplejwt>=5,<6
channels>=4,<5
channels-redis>=4,<5
redis>=5,<6
celery>=5.3,<6
psycopg[binary]>=3.1,<4
python-dotenv>=1,<2
bleach>=6,<7
django-cors-headers>=4,<5
drf-spectacular>=0.27,<0.28
django-storages>=1.14,<2
boto3>=1.28,<2